feat: hashtag wrapper

+++ wrappers/category/mock.go
@@ -3,10 +3,10 @@ package category
 import "go.elastic.co/apm"
 
 type CategoryWrapperMock struct {
-	GetCategoryInternalFn func(categoryIds []int64, limit int64, offset int64, apmTransaction *apm.Transaction, forceLog bool) chan CategoryGetInternalResponseChan
+	GetCategoryInternalFn func(categoryIds []int64, limit int, offset int, apmTransaction *apm.Transaction, forceLog bool) chan CategoryGetInternalResponseChan
 }
 
-func (w *CategoryWrapperMock) GetCategoryInternal(categoryIds []int64, limit int64, offset int64, apmTransaction *apm.Transaction, forceLog bool) chan CategoryGetInternalResponseChan {
+func (w *CategoryWrapperMock) GetCategoryInternal(categoryIds []int64, limit int, offset int, apmTransaction *apm.Transaction, forceLog bool) chan CategoryGetInternalResponseChan {
 	return w.GetCategoryInternalFn(categoryIds, limit, offset, apmTransaction, forceLog)
 }
